refactor(engagement): extract API layer into dojo/engagement/api/ [engagement Phase 6,7,8,9]
