+++ dojo/urls.py
@@ -22,8 +22,6 @@ from dojo.api_v2.views import (
     EndpointMetaImporterView,
     EndpointStatusViewSet,
     EndPointViewSet,
-    EngagementPresetsViewset,
-    EngagementViewSet,
     FindingTemplatesViewSet,
     FindingViewSet,
     ImportLanguagesView,
@@ -60,6 +58,7 @@ from dojo.benchmark.urls import urlpatterns as benchmark_urls
 from dojo.components.urls import urlpatterns as component_urls
 from dojo.development_environment.urls import urlpatterns as dev_env_urls
 from dojo.endpoint.urls import urlpatterns as endpoint_urls
+from dojo.engagement.api.urls import add_engagement_urls
 from dojo.engagement.ui.urls import urlpatterns as eng_urls
 from dojo.finding.urls import urlpatterns as finding_urls
 from dojo.finding_group.urls import urlpatterns as finding_group_urls
@@ -111,8 +110,6 @@ v2_api.register(r"development_environments", DevelopmentEnvironmentViewSet, base
 # RBAC endpoints moved to Pro under legacy authorization:
 #   dojo_groups, dojo_group_members → pro/groups, pro/group_members
 v2_api.register(r"endpoint_meta_import", EndpointMetaImporterView, basename="endpointmetaimport")
-v2_api.register(r"engagements", EngagementViewSet, basename="engagement")
-v2_api.register(r"engagement_presets", EngagementPresetsViewset, basename="engagement_presets")
 v2_api.register(r"finding_templates", FindingTemplatesViewSet, basename="finding_template")
 v2_api.register(r"findings", FindingViewSet, basename="finding")
 # RBAC endpoint moved to Pro under legacy authorization: global_roles → pro/global_roles
@@ -135,6 +132,7 @@ v2_api.register(r"product_api_scan_configurations", ProductAPIScanConfigurationV
 # RBAC endpoints moved to Pro under legacy authorization:
 #   product_groups, product_members → pro/product_groups, pro/product_members
 v2_api = add_product_type_urls(v2_api)
+v2_api = add_engagement_urls(v2_api)
 # RBAC endpoints moved to Pro under legacy authorization:
 #   product_type_members, product_type_groups → pro/product_type_members, pro/product_type_groups
 v2_api.register(r"regulations", RegulationsViewSet, basename="regulations")
