+++ unittests/test_rest_framework.py
@@ -44,7 +44,6 @@ from dojo.api_v2.views import (
     DevelopmentEnvironmentViewSet,
     EndpointStatusViewSet,
     EndPointViewSet,
-    EngagementViewSet,
     FindingTemplatesViewSet,
     FindingViewSet,
     ImportLanguagesView,
@@ -71,6 +70,7 @@ from dojo.asset.api.views import (
     AssetViewSet,
 )
 from dojo.authorization.roles_permissions import Permissions, permission_to_action
+from dojo.engagement.api.views import EngagementViewSet
 from dojo.location.api.endpoint_compat import V3EndpointCompatibleViewSet, V3EndpointStatusCompatibleViewSet
 from dojo.location.api.views import LocationFindingReferenceViewSet, LocationProductReferenceViewSet, LocationViewSet
 from dojo.location.models import Location, LocationFindingReference, LocationProductReference
